requirements.txt
discord-py
git+https://github.com/dolfies/discord.py-self@renamed#egg=selfcord.py[voice]